Create new js file for add and remove functionality
