+++ README.md
@@ -24,7 +24,7 @@ To get a local copy up and running follow these simple example steps.
 
 ### Usage
 
---git clone git@github.com:Addisu87/webpack-demo.git
+-- git clone git@github.com:Addisu87/webpack-demo.git
 
 ### Run tests - npm run build then npm start
