+++ src/index.js
@@ -1,79 +1,5 @@
 import './style.css';
-
-const DisplayTodos = () => {
-  let todos = JSON.parse(localStorage.getItem('todos')) || [];
-  const todoList = document.querySelector('#todo-list');
-  todoList.innerHTML = '';
-
-  todos.forEach((todo) => {
-    const todoItem = document.createElement('div');
-    todoItem.classList.add('todo-item');
-
-    const label = document.createElement('label');
-    const input = document.createElement('input');
-    const span = document.createElement('span');
-    const content = document.createElement('div');
-    const actions = document.createElement('div');
-    const edit = document.createElement('button');
-    const deleteButton = document.createElement('button');
-
-    input.type = 'checkbox';
-    input.checked = todo.done;
-    span.classList.add('bubble');
-    content.classList.add('todo-content');
-    actions.classList.add('actions');
-    edit.classList.add('edit');
-    deleteButton.classList.add('delete');
-
-    content.innerHTML = `<input type="text" value="${todo.content}" readonly>`;
-    edit.innerHTML = 'Edit';
-    deleteButton.innerHTML = 'Delete';
-
-    label.appendChild(input);
-    label.appendChild(span);
-    actions.appendChild(edit);
-    actions.appendChild(deleteButton);
-    todoItem.appendChild(label);
-    todoItem.appendChild(content);
-    todoItem.appendChild(actions);
-
-    todoList.appendChild(todoItem);
-
-    if (todo.done) {
-      todoItem.classList.add('done');
-    }
-
-    input.addEventListener('change', (e) => {
-      todo.done = e.target.checked;
-      localStorage.setItem('todos', JSON.stringify(todos));
-
-      if (todo.done) {
-        todoItem.classList.add('done');
-      } else {
-        todoItem.classList.remove('done');
-      }
-      DisplayTodos();
-    });
-
-    edit.addEventListener('click', () => {
-      const input = content.querySelector('input');
-      input.removeAttribute('readonly');
-      input.focus();
-      input.addEventListener('blur', (e) => {
-        input.setAttribute('readonly', true);
-        todo.content = e.target.value;
-        localStorage.setItem('todos', JSON.stringify(todos));
-        DisplayTodos();
-      });
-    });
-
-    deleteButton.addEventListener('click', () => {
-      todos = todos.filter((task) => task !== todo);
-      localStorage.setItem('todos', JSON.stringify(todos));
-      DisplayTodos();
-    });
-  });
-};
+import DisplayTodos from './displayTodos.js';
 
 window.addEventListener('load', () => {
   const todos = JSON.parse(localStorage.getItem('todos')) || [];
