+++ src/style.css
@@ -1,5 +1,5 @@
 :root {
-  --primary: #868605;
+  --primary: #3d3d29;
   --business: #3a82ee;
   --personal: var(--primary);
   --light: #eee;
